Module Search.Dictionary

Maps from keys to values

type order =

| LT

| EQ

| GT

order represents the concepts of strictly-less-than, equal-to, and strictly greater-than.

module type Comparable = sig ... end

A Comparable is a value that can be compared. The comparison is a total order on the values.

module type Stringable = sig ... end

A Stringable is a value that can be converted to a string.

module type KeySig = sig ... end

A module that matches KeySig is suitable for use as the type of keys in a Dictionary. Keys must be comparable and stringable.

module type ValueSig = sig ... end

A module that matches ValueSig is suitable for use as the type of values in a Dictionary. Values must be stringable.

module type Dictionary = sig ... end

A Dictionary maps keys to values.

module type DictionaryMaker = functor (K : KeySig) -> functor (V : ValueSig) -> Dictionary with module Key = K and module Value = V

A DictionaryMaker is a functor that makes a Dictionary out of modules representing keys and values.