Module type Dictionary.Comparable

A Comparable is a value that can be compared. The comparison is a total order on the values.

type t

The type of comparable values.

val compare : t -> t -> order

compare t1 t2 is LT if t1 is less than t2, EQ if t1 is equal to t2, or GT if t1 is greater than t2.