Module Search.Engine

Search engine.

module type E = sig ... end

An Engine.E indexes words found in text files and answers queries about which files contain which words.

module Make (S : DictionarySet.Set with type Elt.t = string) (D : Dictionary.Dictionary with type Key.t = string and type Value.t = S.t) : E

Make (S) (D) makes an Engine.E out of a Set and a Dictionary. The dictionary D is used to map from words (i.e., strings) to sets of filenames. The set S is used as the type of values in the dictionary (i.e., the sets of filenames).