Module ListDictionary.Make

Make makes a Dictionary implemented with association lists.

Parameters

module K : Dictionary.KeySig

module V : Dictionary.ValueSig

Signature

module Key = K

Key is a module representing the type of keys in the dictionary and functions on them.

module Value = V

Value is a module representing the type of values in the dictionary and functions on them.

type key = Key.t

key is the type of keys in the dictionary.

type value = Value.t

value is the type of values in the dictionary.

type t

t is the type of dictionaries.

val empty : t

empty is the empty dictionary

val is_empty : t -> bool

is_empty d is true iff d is empty.

val size : t -> int

size d is the number of bindings in d. size empty is 0.

val insert : key -> value -> t -> t

insert k v d is d with k bound to v. If k was already bound, its previous value is replaced with v.

val member : key -> t -> bool

member k d is true iff k is bound in d.

val find : key -> t -> value option

find k d is Some v if k is bound to v in d; or if k is not bound, then it is None.

val remove : key -> t -> t

remove k d contains all the bindings of d except a binding for k. If k is not bound in d, then remove returns a dictionary with the same bindings as d.

val fold : ( key -> value -> 'acc -> 'acc ) -> 'acc -> t -> 'acc

fold f init d is f kn vn (f ... (f k1 v1 init) ...), if d binds ki to vi. Bindings are processed in order from least to greatest, where k1 is the least key and kn is the greatest.

val to_list : t -> (key * value) list

to_list d is an association list containing the same bindings as d. The elements in the list are in order from the least key to the greatest. There are no duplicate keys in the list.

include Dictionary.Stringable with type t := t

val to_string : t -> string

to_string v is a textual representation of v.